refactor: techStackDropdown, PositionDropdown, TitleSearch -> 'search'로 분리
